Better organized routing for accounts and posts
